requirements.txt
streamlit
speechrecognition
pyttsx3
pyaudio
openai
langchain
langchain-community
langchain-openai
chromadb
pypdf
faiss-cpu
python-dotenv
